+++ gene2phenotype_project/gene2phenotype_app/views/__init__.py
@@ -1,0 +1,23 @@
+from .base import BaseView, BaseAdd
+
+from .panel import PanelList, PanelDetail, PanelRecordsSummary
+
+from .locus import LocusGene, LocusGeneSummary, GeneFunction
+
+from .disease import GeneDiseaseView, DiseaseDetail, DiseaseSummary, AddDisease
+
+from .curation import (AddCurationData, ListCurationEntries, CurationDataDetail,
+                       UpdateCurationData, PublishRecord)
+
+from .search import SearchView
+
+from .attrib import AttribTypeList, AttribList
+
+from .user import UserList
+
+from .publication import PublicationDetail, AddPublication
+
+from .locus_genotype_disease import (ListMolecularMechanisms, VariantTypesList,
+                                     LocusGenotypeDiseaseDetail, LocusGenotypeDiseaseAddPanel)
+
+from .phenotype import AddPhenotype
