+++ gene2phenotype_project/gene2phenotype_app/views/attrib.py
@@ -1,0 +1,33 @@
+from rest_framework import generics
+from rest_framework.response import Response
+
+from gene2phenotype_app.serializers import AttribTypeSerializer, AttribSerializer
+
+from gene2phenotype_app.models import AttribType, Attrib
+
+
+class AttribTypeList(generics.ListAPIView):
+    queryset = AttribType.objects.all()
+
+    def list(self, request, *args, **kwargs):
+        queryset = self.get_queryset()
+        result = {}
+        for attrib_type in queryset:
+            serializer = AttribTypeSerializer(attrib_type)
+            all_attribs = serializer.get_all_attribs(attrib_type.id)
+            result[attrib_type.code] = all_attribs
+
+        return Response(result)
+
+class AttribList(generics.ListAPIView):
+    lookup_field = 'type'
+    serializer_class = AttribSerializer
+
+    def get_queryset(self):
+        code = self.kwargs['code']
+        return Attrib.objects.filter(type=AttribType.objects.get(code=code))
+    
+    def list(self, request, *args, **kwargs):
+        queryset = self.get_queryset()
+        code_list = [attrib.value for attrib in queryset]
+        return Response({'results':code_list, 'count':len(code_list)})
